feature(node): adding a module system
Adding a module system to allow the extension of bitcore-node with new currencies via npm install
and a config entry
