+++ packages/bitcore-node/src/config.ts
@@ -61,6 +61,7 @@ const Config = function(): ConfigType {
     dbPass: process.env.DB_PASS || '',
     numWorkers: cpus().length,
     chains: {},
+    modules: ['./bitcoin', './bitcoin-cash'],
     services: {
       api: {
         rateLimiter: {
