+++ packages/bitcore-node/src/workers/api.ts
@@ -4,6 +4,7 @@ import { Api } from '../services/api';
 import { Event } from '../services/event';
 import cluster = require('cluster');
 import parseArgv from '../utils/parseArgv';
+import { Modules } from "../modules";
 import '../utils/polyfills';
 require('heapdump');
 
@@ -28,6 +29,9 @@ export const ClusteredApiWorker = async () => {
   } else {
     services.push(Api);
   }
+
+  Modules.loadConfigured();
+
   for (const service of services) {
     await service.start();
   }
