+++ packages/bitcore-node/src/workers/p2p.ts
@@ -1,6 +1,7 @@
 import { P2P } from '../services/p2p';
 import { Storage } from '../services/storage';
 import { Event } from '../services/event';
+import { Modules } from "../modules";
 import '../utils/polyfills';
 require('heapdump');
 const services: Array<any> = [];
@@ -14,6 +15,9 @@ export const P2pWorker = async () => {
   process.on('SIGINT', stop);
 
   services.push(Storage, Event, P2P);
+
+  Modules.loadConfigured();
+
   for (const service of services) {
     await service.start();
   }
