+++ packages/bitcore-node/test/verification/resync.ts
@@ -1,6 +1,6 @@
-import { P2pWorker } from '../../src/services/p2p';
 import { Config } from '../../src/services/config';
 import { Storage } from '../../src/services/storage';
+import { BitcoinP2PWorker } from '../../src/modules/bitcoin/p2p';
 
 (async () => {
   const { CHAIN: chain, NETWORK: network, START, END } = process.env;
@@ -10,7 +10,7 @@ import { Storage } from '../../src/services/storage';
   } else {
     await Storage.start();
     const chainConfig = Config.chainConfig({ chain, network });
-    const worker = new P2pWorker({ chain, network, chainConfig });
+    const worker = new BitcoinP2PWorker({ chain, network, chainConfig });
     await worker.connect();
 
     await worker.resync(Number(START), Number(END));
